move tests, add len() and is_empty()
